feat(context): add Context and custom hook to access cities data
