+++ src/components/CountryList.jsx
@@ -1,8 +1,10 @@
 import Loader from "./Loader";
 import CountryItem from "./CountryItem";
 import Message from "./Message";
+import useCities from "../hooks/useCities";
 
-const CountryList = ({ cities, isLoading }) => {
+const CountryList = () => {
+  const { cities, isLoading } = useCities();
   if (isLoading) return <Loader />;
   if (!cities.length)
     return (
